[MNG-5227] The 'optional' flag of a dependency should be manageable
Updated to add missing management of the optional flag to the ModelBuilder.
The Maven Resolver has been managing the optional flag for years. This just
adds the missing parts to the ModelBuilder to make optional flag management
consistent when building the effective model no matter if resolved or if
built.